+++ freebot_g474/config/config_freebot_g474_motor.cpp
@@ -1,0 +1,82 @@
+Communication System::communication_ = {config::usb1};
+void usb_interrupt() {
+    config::usb1.interrupt();
+}
+Actuator System::actuator_ = {config::fast_loop, config::main_loop, param->startup_param};
+
+float v3v3 = 3.3;
+
+int32_t index_mod = 0;
+
+void system_init() {
+    if (config::motor_encoder.init()) {
+        System::log("Motor encoder init success");
+    } else {
+        System::log("Motor encoder init failure");
+    }
+    if (config::output_encoder.init()) {
+        System::log("Output encoder init success");
+    } else {
+        System::log("Output encoder init failure");
+    }
+    if (drv_regs_error) {
+        System::log("drv configure failure");
+    } else {
+        System::log("drv configure success");
+    }
+    config::torque_sensor.init();
+
+    System::api.add_api_variable("3v3", new APIFloat(&v3v3));
+    std::function<float()> get_t = std::bind(&TempSensor::get_value, &config::temp_sensor);
+    std::function<void(float)> set_t = std::bind(&TempSensor::set_value, &config::temp_sensor, std::placeholders::_1);
+    System::api.add_api_variable("T", new APICallbackFloat(get_t, set_t));
+    System::api.add_api_variable("index_mod", new APIInt32(&index_mod));
+    System::api.add_api_variable("drv_err", new APICallbackUint32(get_drv_status, drv_reset));
+
+    for (auto regs : std::vector<ADC_TypeDef*>{ADC1, ADC3, ADC4, ADC5}) {
+        regs->CR = ADC_CR_ADVREGEN;
+        ns_delay(20000);
+        regs->CR |= ADC_CR_ADCAL;
+        while(regs->CR & ADC_CR_ADCAL);
+        ns_delay(100);
+
+        regs->ISR = ADC_ISR_ADRDY;
+        regs->CR |= ADC_CR_ADEN;
+    }
+    while(!(ADC1->ISR & ADC_ISR_ADRDY));
+    ADC1->CR |= ADC_CR_JADSTART;
+    while(ADC1->CR & ADC_CR_JADSTART);
+
+    v3v3 =  *((uint16_t *) (0x1FFF75AA)) * 3.0 / V_REF_DR;
+    System::log("3v3: " + std::to_string(v3v3));
+
+    config_init();
+
+    ADC1->GCOMP = v3v3*4096;
+    ADC1->CFGR2 |= ADC_CFGR2_GCOMP;
+    ADC1->CR |= ADC_CR_ADSTART;
+    ADC5->CR |= ADC_CR_JADSTART;
+    ADC5->IER |= ADC_IER_JEOCIE;
+    ADC4->CR |= ADC_CR_JADSTART;
+    ADC3->CR |= ADC_CR_JADSTART;
+   
+    TIM1->CR1 = TIM_CR1_CEN; // start main loop interrupt
+    config::usb1.connect();
+    HRTIM1->sMasterRegs.MCR = HRTIM_MCR_TDCEN + HRTIM_MCR_TECEN + HRTIM_MCR_TFCEN; // start high res timer
+}
+
+FrequencyLimiter temp_rate = {10};
+
+void system_maintenance() {
+    if (temp_rate.run()) {
+        ADC1->CR |= ADC_CR_JADSTART;
+        while(ADC1->CR & ADC_CR_JADSTART);
+        config::temp_sensor.read();
+        v3v3 =  *((uint16_t *) (0x1FFF75AA)) * 3.0 * ADC1->GCOMP / 4096.0 / ADC1->JDR2;
+    }
+    index_mod = config::motor_encoder.index_error(param->fast_loop_param.motor_encoder.cpr);
+    config_maintenance();
+}
+
+
+#include "../../motorlib/system.cpp"
